Minor fixes, add kismet

+++ vagrant/create.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+DESTROY=true
 HALT=true
 #HALT=false
 
@@ -12,13 +13,18 @@ fi
 
 if [ "$OPTION" == "vmware" ]; then
     echo "VMware"
-    vagrant destroy vmware_vm --force
-    echo "$D Start vmware_vm " | tee -a date.log 
+    if [ "$DESTROY" = true ] ; then
+        vagrant destroy vmware_vm --force
+    fi
+    D=`date`
+    echo "$D Start vmware_vm " | tee -a vmware_vm.log 
     vagrant up vmware_vm 
     D=`date`
-    echo "$D Finish vmware_vm " | tee -a date.log 
+    echo "$D Finish vmware_vm " | tee -a vmware_vm.log 
     
     # Configure background, etc
+    vagrant halt vmware_vm
+    vagrant up vmware_vm
     timeout 30s vagrant ssh vmware_vm
     if [ "$HALT" = true ] ; then
         vagrant halt vmware_vm
@@ -26,11 +32,17 @@ if [ "$OPTION" == "vmware" ]; then
 
 elif [ $OPTION == "virtualbox" ]; then
     echo "VirtualBox"
-    vagrant destroy virtualbox_vm --force
-    echo "$D Start virtualbox_vm " | tee -a date.log 
+    if [ "$DESTROY" = true ] ; then
+        vagrant destroy virtualbox_vm --force
+    fi
+    D=`date`
+    echo "$D Start virtualbox_vm " | tee -a virtualbox_vm.log
     vagrant up virtualbox_vm 
-    echo "$D Finish virtualbox_vm " | tee -a date.log 
+    D=`date`
+    echo "$D Finish virtualbox_vm " | tee -a virtualbox_vm.log
     # Configure background, etc
+    vagrant halt virtualbox_vm
+    vagrant up virtualbox_vm
     timeout 30s vagrant ssh virtualbox_vm
     if [ "$HALT" = true ] ; then
         vagrant halt virtualbox_vm
